+++ src/helper.ts
@@ -2,16 +2,19 @@
 
 import { parse } from '@babel/parser';
 
-export function getStyles(text: string) {
+type StyleDetail = {
+  rootName: string;
+  styles: any;
+  location: any;
+  styleType: 'normal' | 'arrow';
+};
+export function getStyles(text: string): StyleDetail[] {
   const ast = parse(text, {
     sourceType: 'unambiguous',
     plugins: ['jsx', 'classProperties', 'typescript'],
   });
 
-  const styles: any = {};
-  let globalStyleName = '';
-  let stylesType = 'normal';
-  let styleLocation = {}
+  const styleList: StyleDetail[] = [];
 
   ast.program.body
     .filter((node) => node.type === 'VariableDeclaration')
@@ -24,10 +27,15 @@ export function getStyles(text: string) {
           const obj = callee?.object;
           const property = callee?.property;
 
-          if (init.type === 'CallExpression' && obj?.name === 'StyleSheet' && property?.name === 'create' && !globalStyleName) {
-            globalStyleName = item.id.name;
+          const styles = {};
+          let rootName = '';
+          let styleType = 'normal';
+          let location = {};
 
-            styleLocation = item.loc;
+          if (init.type === 'CallExpression' && obj?.name === 'StyleSheet' && property?.name === 'create') {
+            rootName = item.id.name;
+
+            location = item.loc;
             init?.arguments[0].properties.forEach((item) => {
               const name = item.key.name;
               styles[name] = {
@@ -35,10 +43,17 @@ export function getStyles(text: string) {
                 details: { item },
               };
             });
-          } else if (init.type === 'ArrowFunctionExpression' && init.body.callee?.object.name === 'StyleSheet' && !globalStyleName) {
-            globalStyleName = item.id.name ?? '';
-            stylesType = 'arrow';
-            styleLocation = item.loc;
+
+            styleList.push({
+              styles,
+              rootName,
+              styleType,
+              location,
+            });
+          } else if (init.type === 'ArrowFunctionExpression' && init.body.callee?.object.name === 'StyleSheet') {
+            rootName = item.id.name ?? '';
+            styleType = 'arrow';
+            location = item.loc;
 
             init.body.arguments[0].properties.forEach((item) => {
               const name = item.key.name;
@@ -47,25 +62,31 @@ export function getStyles(text: string) {
                 details: { item },
               };
             });
+
+            styleList.push({
+              styles,
+              rootName,
+              styleType,
+              location,
+            });
           }
         });
     });
-
-  for (let item in styles) {
-    const name = item;
-    let styleToMatch = `${globalStyleName}.${name}`;
-    if (stylesType === 'arrow') {
-      styleToMatch = `${globalStyleName}(.+).${name}`;
+    
+  for (let i = 0; i < styleList.length; i++) {
+    const styleObj = styleList[i]
+    for (let item in styleObj.styles) {
+      const name = item;
+      let styleToMatch = `${styleObj.rootName}.${name}`;
+      if (styleObj.styleType === 'arrow') {
+        styleToMatch = `${styleObj.rootName}(.+).${name}`;
+      }
+      const regex = new RegExp(styleToMatch, 'g');
+      const matches = text.match(regex);
+      const useCount = matches ? matches.length : 0;
+      styleList[i].styles[name].usage = useCount;
     }
-    const regex = new RegExp(styleToMatch, 'g');
-    const matches = text.match(regex);
-    const useCount = matches ? matches.length : 0;
-    styles[name].usage = useCount;
   }
 
-  return {
-    styles,
-    globalStyleName,
-    styleLocation
-  };
+  return styleList;
 }