+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change Log
 
+## [0.0.5] - 2023-08-18
+- Added support for multiple styles created in the same file
+- Added highlight for unused style
+- Added refetch styles when file is saved
+- Added display for style object name
+
 ## [0.0.4] - 2023-08-17
 - Added jump to style feature